refactor: feature:camera module package
