--- feature/camera/api/src/main/java/com/teamyg/camera/api/NavKeyCameraSystem.kt
@@ -1,7 +1,0 @@
-package com.teamyg.camera.api
-
-import androidx.navigation3.runtime.NavKey
-import kotlinx.serialization.Serializable
-
-@Serializable
-data object NavKeyCameraSystem : NavKey
