+++ feature/camera/api/src/main/java/com/teamyg/parfait/feature/camera/api/NavKeyCameraCustom.kt
@@ -1,0 +1,7 @@
+package com.teamyg.parfait.feature.camera.api
+
+import androidx.navigation3.runtime.NavKey
+import kotlinx.serialization.Serializable
+
+@Serializable
+data object NavKeyCameraCustom : NavKey
