--- feature/camera/impl/src/main/java/com/teamyg/camera/impl/component/controls/FlipCameraButton.kt
@@ -1,45 +1,0 @@
-package com.teamyg.camera.impl.component.controls
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import com.tjyg.core.ui.preview.PreviewBox
-import com.tjyg.core.ui.preview.YGPreview
-
-@Composable
-internal fun FlipCameraButton(
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    IconButton(
-        onClick = onClick,
-        modifier = modifier
-            .size(48.dp)
-            .clip(CircleShape)
-            .background(Color(0xFF333333)),
-    ) {
-        Text(
-            text = "\uD83D\uDD04",
-            color = Color.White,
-            fontSize = 24.sp,
-            fontWeight = FontWeight.Bold,
-        )
-    }
-}
-
-@YGPreview
-@Composable
-private fun PreviewFlipCameraButton() = PreviewBox {
-    FlipCameraButton(
-        onClick = {},
-    )
-}
