--- feature/camera/impl/src/main/java/com/teamyg/camera/impl/navigation/NavigationModule.kt
@@ -1,20 +1,0 @@
-package com.teamyg.camera.impl.navigation
-
-import androidx.navigation3.runtime.EntryProviderScope
-import androidx.navigation3.runtime.NavKey
-import com.teamyg.navigation.Navigator
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ActivityRetainedComponent
-import dagger.multibindings.IntoSet
-
-@Module
-@InstallIn(ActivityRetainedComponent::class)
-object NavigationModule {
-    @IntoSet
-    @Provides
-    fun provideFeatureCameraEntryBuilder(): EntryProviderScope<NavKey>.(Navigator) -> Unit = {
-        featureCameraEntryBuilder(navigator = it)
-    }
-}
