+++ feature/camera/impl/src/main/java/com/teamyg/parfait/feature/camera/impl/util/CameraFileProvider.kt
@@ -1,0 +1,50 @@
+package com.teamyg.parfait.feature.camera.impl.util
+
+import android.content.Context
+import android.net.Uri
+import androidx.core.content.FileProvider
+import java.io.File
+import kotlin.time.Clock
+import kotlinx.datetime.LocalDateTime
+import kotlinx.datetime.TimeZone
+import kotlinx.datetime.format
+import kotlinx.datetime.format.FormatStringsInDatetimeFormats
+import kotlinx.datetime.format.byUnicodePattern
+import kotlinx.datetime.toLocalDateTime
+
+internal object CameraFileProvider {
+    private const val AUTHORITY_SUFFIX = ".camera.fileprovider"
+    private const val CAMERA_DIR_NAME = "camera"
+    private const val FILE_NAME_PATTERN = "yyyyMMdd_HHmmss"
+
+    @OptIn(FormatStringsInDatetimeFormats::class)
+    fun createImageFile(context: Context): File {
+        val cameraDir = File(
+            context.cacheDir,
+            CAMERA_DIR_NAME,
+        ).apply {
+            mkdirs()
+        }
+
+        val timestamp = Clock.System
+            .now()
+            .toLocalDateTime(TimeZone.currentSystemDefault())
+            .format(LocalDateTime.Format { byUnicodePattern(FILE_NAME_PATTERN) })
+
+        return File(cameraDir, "IMG_$timestamp.jpg")
+    }
+
+    fun toContentUri(
+        context: Context,
+        file: File,
+    ): Uri = FileProvider.getUriForFile(
+        context,
+        context.packageName + AUTHORITY_SUFFIX,
+        file,
+    )
+
+    fun createImageUri(context: Context): Uri = toContentUri(
+        context = context,
+        file = createImageFile(context),
+    )
+}
